requirements.txt
apipkg==1.5
appdirs==1.4.3
Arpeggio==1.9.2
asgiref==3.3.1
asynctest==0.13.0
attrs==19.3.0
beautifulsoup4==4.8.2
bitstring==3.1.5
blinker==1.4
Brotli==1.0.7
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.1
colorama==0.4.3
coverage==5.1
cryptography==3.2.1
distlib==0.3.0
dnspython==1.16.0
docutils==0.16
entrypoints==0.3
eventlet==0.29.1
execnet==1.7.1
filelock==3.0.12
flake8==3.7.9
Flask==1.1.2
Flask-SocketIO==4.3.2
Flask-SQLAlchemy==2.4.4
flatten-json==0.1.6
greenlet==0.4.17
h11==0.9.0
h2==4.0.0
hpack==4.0.0
hyperframe==6.0.0
idna==2.9
importlib-metadata==1.6.0
importlib-resources==1.5.0
IP2Location==8.4.1
itsdangerous==1.1.0
Jinja2==2.11.1
kaitaistruct==0.8
ldap3==2.8.1
MarkupSafe==1.1.1
mccabe==0.6.1
mitmproxy==5.3.0
more-itertools==8.2.0
msgpack==1.0.2
mypy==0.770
mypy-extensions==0.4.3
mysqlclient==2.0.1
packaging==20.3
parver==0.3.0
passlib==1.7.2
pluggy==0.13.1
protobuf==3.13.0
psutil==5.8.0
publicsuffix2==2.20191221
py==1.8.1
pyasn1==0.4.8
pycodestyle==2.5.0
pycparser==2.20
pydevd==1.9.0
pydevd-pycharm==201.6668.115
pyflakes==2.1.1
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyperclip==1.8.0
pytest==5.4.1
pytest-asyncio==0.10.0
pytest-cov==2.8.1
pytest-forked==1.1.3
pytest-timeout==1.3.4
pytest-xdist==1.31.0
python-dotenv==0.15.0
python-engineio==3.14.2
python-socketio==4.6.1
requests==2.23.0
rstcheck==3.3.1
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
six==1.14.0
sortedcontainers==2.2.2
soupsieve==2.0
SQLAlchemy==1.3.20
tinycss==0.4
toml==0.10.0
tornado==6.0.4
tox==3.14.6
typed-ast==1.4.1
typing-extensions==3.7.4.2
urllib3==1.25.8
urwid==2.1.2
virtualenv==20.0.17
wcwidth==0.1.9
Werkzeug==1.0.1
wsproto==0.15.0
Yapsy==1.11.223
yara-python==3.11.0
zipp==3.1.0
zstandard==0.13.0
